Add files via upload
This is my first hands-on project with Python. The famous Titanic database is used. I have used various libraries like SeaBorn, Pandas, Matplotlib, etc. to make inferences about the database.

Various graphs in this project help us understand the power of visualization that is provided by different libraries in Python and how we can interpret a large dataset in a better manner with the help of such techniques.